ci: Remove artifact attestation step from workflow (#425)
<!-- Please use this template for your pull request. -->
<!-- Please use the sections that you need and delete other sections -->

## This PR
<!-- add the description of the PR here -->

This pull request includes a small change to the
`.github/workflows/ci.yml` file. The change removes the step that
generates artifact attestation using the
`actions/attest-build-provenance` action.

* Removed the `Generate artifact attestation` step from the CI workflow
in `.github/workflows/ci.yml`.

### Related Issues
<!-- add here the GitHub issue that this PR resolves if applicable -->

Fixes #423 

### Notes
<!-- any additional notes for this PR -->

This only removes execution from the PRs and forks. It will not be
removed from the main publishing action.

Signed-off-by: André Silva <2493377+askpt@users.noreply.github.com>