test: Add E2E Steps for contextMerging.feature tests (#422)
<!-- Please use this template for your pull request. -->
<!-- Please use the sections that you need and delete other sections -->

## This PR

- Add new Reqnroll steps for contextMerging.feature ahead of merging
#395

### Related Issues
<!-- add here the GitHub issue that this PR resolves if applicable -->

Fixes #399

### Notes
<!-- any additional notes for this PR -->

I was following the [requirement
3.2.3](https://openfeature.dev/specification/sections/evaluation-context#requirement-323)

I'm assuming with these additions that "Before Hook" is the same as
"Invocation" in the dotnet-sdk. If this is not the case then please let
me know!

In order to fetch the merged context I use a test hook with a function
that can set the EvaluationContext on the state. There are probably
better abstractions to use

I haven't updated the spec submodule in these changes either, although I
can add that if needed

Let me know if you have any concerns or feedback 🏗️

### Follow-up Tasks
<!-- anything that is related to this PR but not done here should be
noted under this section -->
<!-- if there is a need for a new issue, please link it here -->

### How to test
<!-- if applicable, add testing instructions under this section -->

---------

Signed-off-by: Kyle Julian <38759683+kylejuliandev@users.noreply.github.com>
Co-authored-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>
Co-authored-by: André Silva <2493377+askpt@users.noreply.github.com>