feat: Add support for hook data. (#387)
<!-- Please use this template for your pull request. -->
<!-- Please use the sections that you need and delete other sections -->

## This PR

Adds support for hook data.
https://github.com/open-feature/spec/pull/273

### Related Issues
<!-- add here the GitHub issue that this PR resolves if applicable -->

### Notes
<!-- any additional notes for this PR -->

I realized that the 4.6.1 section of the spec wasn't consistent with the
expected usage. Basically it over-specifies the typing of the hook data
matching that of the evaluation context. That is one possible approach,
it would just mean a bit more work on the part of the hook implementers.

In the earlier example in the spec I put a `Span` in the hook data:
```
  public Optional<EvaluationContext> before(HookContext context, HookHints hints) {
    SpanBuilder builder = tracer.spanBuilder('sample')
    .setParent(Context.current().with(Span.current()));
    Span span = builder.startSpan()
    context.hookData.set("span", span);
  }
  public void after(HookContext context, FlagEvaluationDetails details, HookHints hints) {
    // Only accessible by this hook for this specific evaluation.
    Object value = context.hookData.get("span");
    if (value instanceof Span) {
      Span span = (Span) value;
      span.end();
    }
  }
```

This is only possible if the hook data allows specification of any
`object` instead of being limited to the immutable types of a context.
For hooks hook data this is safe because only the hook mutating the data
will have access to that data. Additionally the execution of the hook
will be in sequence with the evaluation (likely in a single thread).

The alternative would be to store data in the hook, and use the hook
data to know when to remove it.
Something like this:
```
  public Optional<EvaluationContext> before(HookContext context, HookHints hints) {
    SpanBuilder builder = tracer.spanBuilder('sample')
    .setParent(Context.current().with(Span.current()));
    Span span = builder.startSpan()

    String storageId = Uuid();
    this.tmpData.set(storageId, span);
    context.hookData.set("span", storageId);
  }
  public void after(HookContext context, FlagEvaluationDetails details, HookHints hints) {
    // Only accessible by this hook for this specific evaluation.
    Object value = context.hookData.get("span");
    if (value) {
      String id = value.AsString();
      Span span= this.tmpData.get(id);
      span.end();
    }
  }
```


### Follow-up Tasks
<!-- anything that is related to this PR but not done here should be
noted under this section -->
<!-- if there is a need for a new issue, please link it here -->

### How to test
<!-- if applicable, add testing instructions under this section -->

---------

Signed-off-by: Ryan Lamb <4955475+kinyoklion@users.noreply.github.com>
Co-authored-by: chrfwow <christian.lutnik@dynatrace.com>