test: Fix flaky integration flaky (#469)
### This PR

Reworked the `VerifyMultipleHandlersAreRegisteredAsync` test assertions
in the Integration tests to make them more reliable.

### Releated issues

Fixes #468 

### Notes

Before we were asserting that the handler was invoked twice. However, we
don't necessarily need to test how many invocations occured but rather
that the invocations happened.

I've tweaked the counter to increment when the handler is injected
(which should always be twice) and added a couple of boolean variables
that we can use to track whether each handler did invoke at least once.
I think this makes for a more accurate and reliable test case.

---------

Signed-off-by: Kyle Julian <38759683+kylejuliandev@users.noreply.github.com>