docs: Add AspNetCore sample app (#477)
<!-- Please use this template for your pull request. -->
<!-- Please use the sections that you need and delete other sections -->

## This PR
<!-- add the description of the PR here -->

- Adds simple a AspNetCore web application with a `/welcome` minimal API
that will show one of two messages based the configuration of the
`InMemory` feature provider.

### Related Issues
<!-- add here the GitHub issue that this PR resolves if applicable -->

### Notes
<!-- any additional notes for this PR -->

While the
[README](https://github.com/open-feature/dotnet-sdk/blob/6c44db9237748735a22a162a88f61b2de7f3e9bd/README.md)
for OpenFeature dotnet is pretty comprehensive, I think it can be quite
intimidating to developers who are unfamilar with feature flagging and
OpenFeature. Providing a basic application that can be cloned locally
and run with minimal setup would improve the onboarding and upskill
experience.

Additionally, as a contributor I find myself frequently creating simple
applications to test out changes to OpenFeature. This would save the
effort of stashing local changes and make it easier to more quickly test
or prototype changes.

The example could be extended with endpoints that explore the other
types of feature flags and how you might utilise them within an
application.

### Follow-up Tasks
<!-- anything that is related to this PR but not done here should be
noted under this section -->
<!-- if there is a need for a new issue, please link it here -->

Update the
[README](https://github.com/open-feature/dotnet-sdk/blob/6c44db9237748735a22a162a88f61b2de7f3e9bd/README.md)
with a reference to the sample application to help guide developers.

### How to test
<!-- if applicable, add testing instructions under this section -->

---------

Signed-off-by: Kyle Julian <38759683+kylejuliandev@users.noreply.github.com>